refactor: extraigo los componentes de Playground en carpetas
