+++ src/frontend/CodePlayground/Editor/EditorButton.tsx
@@ -1,0 +1,15 @@
+import { JSX, MouseEventHandler } from "react";
+
+type Props = {
+  icon: string; // despues pasemoslo a JSX.Element con lucide-icons
+  onClick: MouseEventHandler<HTMLButtonElement> | undefined;
+  title: string;
+};
+
+export default function EditorButton({ title, onClick, icon }: Props) {
+  return (
+    <button onClick={onClick} title={title}>
+      {icon}
+    </button>
+  );
+}
