+++ src/frontend/CodePlayground/Header/PlaygroundHeader.tsx
@@ -1,17 +1,15 @@
-import { CodeIcon, PencilIcon, TerminalIcon } from "../icons/Icons";
-import { usePlayground } from "./PlaygroundContext";
-import { useTranslation } from "react-i18next";
-import { PlaygroundViews } from "../model/common";
-import { Dispatch, JSX, SetStateAction } from "react";
+import { CodeIcon, PencilIcon, TerminalIcon } from "../../icons/Icons";
+import { usePlayground } from "../PlaygroundContext";
+import { PlaygroundViews } from "../../model/common";
+import { PlaygroundButton } from "./PlaygroundButton";
 
 type Props = {};
 
 export default function PlaygroundHeader({}: Props) {
-  const { t } = useTranslation();
   const { exercise, activeView, setActiveView } = usePlayground();
   return (
     <div className="flex items-center border-b">
-      <HeaderButton
+      <PlaygroundButton
         icon={
           <PencilIcon width={15} height={15} className="fill-mumuki-teal" />
         }
@@ -21,7 +19,7 @@ export default function PlaygroundHeader({}: Props) {
         setActiveView={setActiveView}
       />
       {exercise.extra && (
-        <HeaderButton
+        <PlaygroundButton
           icon={
             <CodeIcon width={15} height={15} className="fill-mumuki-teal" />
           }
@@ -31,7 +29,7 @@ export default function PlaygroundHeader({}: Props) {
           setActiveView={setActiveView}
         />
       )}
-      <HeaderButton
+      <PlaygroundButton
         icon={
           <TerminalIcon width={15} height={15} className="fill-mumuki-teal" />
         }
@@ -43,29 +41,3 @@ export default function PlaygroundHeader({}: Props) {
     </div>
   );
 }
-
-type ButtonProps = {
-  translationKey: string;
-  icon: JSX.Element;
-  view: PlaygroundViews;
-  activeView: PlaygroundViews
-  setActiveView: (view: PlaygroundViews) => void;
-};
-
-const HeaderButton = ({
-  translationKey,
-  icon,
-  view,
-  activeView,
-  setActiveView,
-}: ButtonProps) => {
-  const { t } = useTranslation();
-  return (
-    <button
-      onClick={() => setActiveView(view)}
-      className={"flex items-center gap-1 px-2 py-1 rounded hover:bg-gray-50 transition " + (activeView === view ? " border-x border-t bg-gray-100" : "")}>
-      {icon}
-      <span className="text-mumuki-teal">{t(translationKey)}</span>
-    </button>
-  );
-};
